# public/swagger.json
{
    "components": {
        "examples": {},
        "headers": {},
        "parameters": {},
        "requestBodies": {},
        "responses": {},
        "schemas": {
            "IJsonResponseSuccess_any_": {
                "properties": {
                    "status": {
                        "type": "number",
                        "format": "double"
                    },
                    "message": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "data": {}
                },
                "required": [
                    "status"
                ],
                "type": "object",
                "additionalProperties": false
            },
            "IFormCreationParam": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "formsAppEnvironmentId": {
                        "type": "number",
                        "format": "double"
                    },
                    "description": {
                        "type": "string"
                    },
                    "organisationId": {
                        "type": "string"
                    },
                    "elements": {
                        "items": {},
                        "type": "array"
                    },
                    "isAuthenticated": {
                        "type": "boolean"
                    },
                    "submissionEvents": {
                        "items": {},
                        "type": "array"
                    },
                    "postSubmissionAction": {
                        "type": "string"
                    },
                    "formsAppIds": {
                        "items": {
                            "type": "number",
                            "format": "double"
                        },
                        "type": "array"
                    }
                },
                "type": "object",
                "additionalProperties": false
            },
            "IFormUpdateParam": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "formsAppEnvironmentId": {
                        "type": "number",
                        "format": "double"
                    },
                    "description": {
                        "type": "string"
                    },
                    "organisationId": {
                        "type": "string"
                    },
                    "elements": {
                        "items": {},
                        "type": "array"
                    },
                    "isAuthenticated": {
                        "type": "boolean"
                    },
                    "submissionEvents": {
                        "items": {},
                        "type": "array"
                    },
                    "postSubmissionAction": {
                        "type": "string"
                    },
                    "formsAppIds": {
                        "items": {
                            "type": "number",
                            "format": "double"
                        },
                        "type": "array"
                    },
                    "id": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "required": [
                    "id"
                ],
                "type": "object",
                "additionalProperties": false
            }
        },
        "securitySchemes": {}
    },
    "info": {
        "title": "onblink-sdk",
        "version": "1.0.0",
        "license": {
            "name": "ISC"
        },
        "contact": {}
    },
    "openapi": "3.0.0",
    "paths": {
        "/civicplus/forms/search": {
            "get": {
                "operationId": "SearchForms",
                "responses": {
                    "200": {
                        "description": "Return array of forms",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IJsonResponseSuccess_any_"
                                }
                            }
                        }
                    }
                },
                "description": "Search forms",
                "tags": [
                    "Forms"
                ],
                "security": [],
                "parameters": [
                    {
                        "description": "- Optional",
                        "in": "query",
                        "name": "name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/civicplus/forms/{id}": {
            "get": {
                "operationId": "GetForm",
                "responses": {
                    "200": {
                        "description": "Return array of forms",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IJsonResponseSuccess_any_"
                                }
                            }
                        }
                    }
                },
                "description": "Get form by id",
                "tags": [
                    "Forms"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "format": "double",
                            "type": "number"
                        }
                    }
                ]
            }
        },
        "/civicplus/forms/create": {
            "post": {
                "operationId": "CreateForm",
                "responses": {
                    "200": {
                        "description": "Return array of forms",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IJsonResponseSuccess_any_"
                                }
                            }
                        }
                    }
                },
                "description": "Create a new form",
                "tags": [
                    "Forms"
                ],
                "security": [],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IFormCreationParam"
                            }
                        }
                    }
                }
            }
        },
        "/civicplus/forms/update": {
            "put": {
                "operationId": "UpdateForm",
                "responses": {
                    "200": {
                        "description": "Return array of forms",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IJsonResponseSuccess_any_"
                                }
                            }
                        }
                    }
                },
                "description": "Update a new form",
                "tags": [
                    "Forms"
                ],
                "security": [],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/IFormUpdateParam"
                            }
                        }
                    }
                }
            }
        }
    },
    "servers": [
        {
            "url": "/api"
        }
    ]
}